refactor(chat): extract build_chat_prompt into chat_context.py
Move the chat prompt builder (and its mtime/section caches) out of awake.py
into a self-contained chat_context module so the dedicated chat process can
build the identical prompt without importing the bridge loop. soul/summary are
read via bridge_state getters (mtime-cached) so edits land on the next reply
without a restart — avoiding PR #1088's stale-personality defect. awake keeps
the historical private names as re-exports; prompt/cache tests move to
test_chat_context.py. (specs/007-chat-process, FR-004)
